Restore Version Sort
Since we only build the book on Linux for now, restore the `--version-sort` flag for gnu sort. This makes me feel better that when sorting numbering oddities (e.g. multiple digits) will be handled correctly.

This was removed when I was trying to get this script to work on Windows and OSX, which is no longer relevant.