Retry locking the mutex before touching waker list.
Moving the try_lock code to before touching the waker list is
sound, because the waker list can only ever be accessed with `blocked`
hold, so as long as we retry lock it while having `blocked` locked, we
are okay.

This code also set both LOCK and BLOCKED in the same atomic op. This
has some performance improvements by touching the atomic variable 1
less time when inserting the entry.

Signed-off-by: Gary Guo <gary@garyguo.net>